requirements.txt
requests
beautifulsoup4
pandas
numpy
SQLAlchemy
pyodbc
streamlit>=1.30.0
plotly
python-dotenv
yfinance
rapidfuzz